+++ wv-wild-web/src/data/state-parks/watoga/facilities.ts
@@ -1,0 +1,327 @@
+/**
+ * Watoga State Park - Facilities Module
+ * SPEC-18 Phase 4: Lodging, camping, pools, visitor centers, amenities
+ *
+ * Part of watoga-sp.ts modular split (SPEC-24 compliance)
+ * @module data/state-parks/watoga/facilities
+ */
+
+import type {
+  Lodging,
+  Camping,
+  Pool,
+  VisitorCenter,
+  PicnicArea,
+  Playground,
+  OtherAmenity,
+} from '../../../types/state-park-template-types';
+import { WV_STATE_PARKS_RESERVATION_URL } from '../shared-constants';
+
+export const watogaFacilities = {
+  lodging: {
+    lodges: [
+      {
+        name: 'Watoga Lake Lodge',
+        rooms: 30,
+        roomTypes: [
+          'Standard Room (2 double beds)',
+          'King Room',
+          'Accessible Room',
+        ],
+        amenities: [
+          'Full-service restaurant',
+          'Conference center',
+          'Wi-Fi in lodge',
+          'Daily housekeeping',
+          'Mini-refrigerator',
+          'Coffee maker',
+          'Cable TV',
+          'Climate control',
+          'Private bathrooms',
+          'Lake views (select rooms)',
+        ],
+        hasRestaurant: true,
+        hasConferenceFacilities: true,
+        description: 'Historic Watoga Lake Lodge offers comfortable rooms with modern amenities and stunning lake views. On-site restaurant serves breakfast, lunch, and dinner.',
+        bookingUrl: WV_STATE_PARKS_RESERVATION_URL,
+        priceRange: '$95-$150/night',
+        seasonalOperation: 'Year-round',
+      },
+    ],
+    cabins: [
+      {
+        cabinNumber: 'Standard Cabin 1-20',
+        name: 'Standard Cabins',
+        bedrooms: 2,
+        bathrooms: 1,
+        maxOccupancy: 6,
+        kitchenType: 'full' as const,
+        hasFireplace: false,
+        petFriendly: false,
+        accessible: false,
+        seasonalAvailability: 'Year-round',
+        amenities: [
+          'Full kitchen',
+          'Electric heat',
+          'Hot water',
+          'Linens provided',
+          'Cookware',
+          'Private bathroom',
+          'Screened porch',
+        ],
+        hasPorch: true,
+        hasGrill: true,
+        priceRange: '$110-$165/night',
+        bookingUrl: WV_STATE_PARKS_RESERVATION_URL,
+      },
+      {
+        cabinNumber: 'Deluxe Cabin 21-28',
+        name: 'Deluxe Cabins',
+        bedrooms: 3,
+        bathrooms: 2,
+        maxOccupancy: 8,
+        kitchenType: 'full' as const,
+        hasFireplace: true,
+        fireplaceType: 'gas' as const,
+        petFriendly: false,
+        accessible: false,
+        seasonalAvailability: 'Year-round',
+        amenities: [
+          'Full kitchen',
+          'Gas fireplace',
+          'Two bathrooms',
+          'Electric heat',
+          'Hot water',
+          'Linens provided',
+          'Dishwasher',
+          'Microwave',
+          'Large screened porch',
+        ],
+        hasPorch: true,
+        hasGrill: true,
+        priceRange: '$135-$195/night',
+        bookingUrl: WV_STATE_PARKS_RESERVATION_URL,
+      },
+      {
+        cabinNumber: 'Pet-Friendly Cabin 29-31',
+        name: 'Pet-Friendly Cabins',
+        bedrooms: 2,
+        bathrooms: 1,
+        maxOccupancy: 6,
+        kitchenType: 'full' as const,
+        hasFireplace: false,
+        petFriendly: true,
+        accessible: false,
+        seasonalAvailability: 'Year-round',
+        amenities: [
+          'Full kitchen',
+          'Pet-friendly (fee applies)',
+          'Fenced yard area',
+          'Electric heat',
+          'Hot water',
+          'Linens provided',
+        ],
+        hasPorch: true,
+        hasGrill: true,
+        priceRange: '$120-$175/night (plus $25 pet fee)',
+        bookingUrl: WV_STATE_PARKS_RESERVATION_URL,
+        description: 'Limited pet-friendly cabins available. Maximum 2 pets, additional fee applies. Pets must be leashed.',
+      },
+      {
+        cabinNumber: 'Accessible Cabin 32-33',
+        name: 'ADA Accessible Cabins',
+        bedrooms: 2,
+        bathrooms: 1,
+        maxOccupancy: 6,
+        kitchenType: 'full' as const,
+        hasFireplace: false,
+        petFriendly: false,
+        accessible: true,
+        seasonalAvailability: 'Year-round',
+        amenities: [
+          'Wheelchair ramp',
+          'Wider doorways',
+          'Roll-in shower',
+          'Accessible kitchen',
+          'Grab bars',
+          'Lower light switches',
+          'Visual alarms',
+        ],
+        hasPorch: true,
+        hasGrill: true,
+        priceRange: '$110-$165/night',
+        bookingUrl: WV_STATE_PARKS_RESERVATION_URL,
+      },
+    ],
+  } satisfies Lodging,
+
+  camping: {
+    campgrounds: [
+      {
+        name: 'Riverside Campground',
+        siteCount: 88,
+        hookupTypes: ['electric', 'water', 'sewer'] as const,
+        bathhouse: true,
+        dumpStation: true,
+        maxRVLength: '45 feet',
+        season: 'April 1 - October 31',
+        amenities: [
+          'Full hookups available',
+          'Hot showers',
+          'Flush toilets',
+          'Laundry facilities',
+          'Camp store',
+          'Playground',
+          'Picnic tables',
+          'Fire rings',
+        ],
+        accessible: true,
+        accessibleSiteCount: 6,
+        fees: '$28-$38/night (full hookup), $25/night (electric only)',
+        bookingUrl: WV_STATE_PARKS_RESERVATION_URL,
+      },
+    ],
+  } satisfies Camping,
+
+  pools: [
+    {
+      type: 'outdoor' as const,
+      name: 'Watoga Swimming Pool',
+      depth: '3-8 feet',
+      features: [
+        'Olympic-size pool',
+        'Diving board',
+        'Lap lanes',
+        'Shallow wading area',
+        'Sundeck',
+      ],
+      lifeguard: true,
+      seasonalOperation: 'Memorial Day - Labor Day',
+      fees: '$5/person, $15/family',
+      accessible: true,
+      accessibilityFeatures: [
+        'Pool lift',
+        'Accessible changing rooms',
+        'Nearby accessible parking',
+      ],
+    },
+  ] satisfies Pool[],
+
+  visitorCenters: [
+    {
+      name: 'Brooks Memorial Arboretum Nature Center',
+      type: 'nature_center' as const,
+      exhibits: [
+        'Tree species identification displays',
+        'Local forest ecology',
+        'Wildlife dioramas',
+        'Historic CCC exhibits',
+        'Seasonal wildflower displays',
+      ],
+      programs: [
+        'Guided arboretum walks',
+        'Tree identification workshops',
+        'Children\'s nature programs',
+      ],
+      staffed: true,
+      giftShop: false,
+      restrooms: true,
+      accessibility: 'Fully accessible facility',
+      contact: '304-799-4087',
+      admissionFee: 'Free',
+    },
+  ] satisfies VisitorCenter[],
+
+  picnicAreas: [
+    {
+      type: 'pavilion' as const,
+      name: 'Lakeside Pavilion',
+      shelterCount: 1,
+      capacity: 100,
+      reservable: true,
+      grills: true,
+      electricity: true,
+      amenities: [
+        'Lake views',
+        'Tables for 100',
+        'Nearby restrooms',
+        'Playground access',
+        'Beach access',
+      ],
+      accessible: true,
+      reservationUrl: WV_STATE_PARKS_RESERVATION_URL,
+      rentalFee: '$75/day',
+      nearParking: true,
+    },
+  ] satisfies PicnicArea[],
+
+  playgrounds: [
+    {
+      name: 'Campground Playground',
+      ageGroups: ['2-5 years', '5-12 years'],
+      equipment: [
+        'Swing sets',
+        'Slides',
+        'Climbing structure',
+        'Monkey bars',
+        'Spring riders',
+      ],
+      surfaceType: 'rubber' as const,
+      shade: true,
+      accessibility: true,
+      accessibilityFeatures: [
+        'Accessible play equipment',
+        'Transfer platforms',
+        'Sensory play elements',
+      ],
+      nearRestrooms: true,
+      nearParking: true,
+    },
+  ] satisfies Playground[],
+
+  otherAmenities: [
+    {
+      type: 'golf_course' as const,
+      name: 'Watoga Golf Course',
+      description: '18-hole championship course designed by renowned architect. Par 72, 6,200 yards.',
+      fees: '$25-$45 per round (seasonal rates)',
+      seasonal: true,
+      season: 'April - October',
+      details: {
+        holes: '18',
+        par: '72',
+        yardage: '6,200',
+        cartRental: '$15',
+        clubRental: '$10',
+      },
+    },
+    {
+      type: 'equipment_rental' as const,
+      name: 'Boat Rentals',
+      description: 'Paddleboat, kayak, and canoe rentals on Watoga Lake',
+      fees: '$10-$25/hour',
+      seasonal: true,
+      season: 'May - September',
+      details: {
+        paddleboats: '$15/hour',
+        kayaks: '$10/hour',
+        canoes: '$12/hour',
+      },
+    },
+    {
+      type: 'gift_shop' as const,
+      name: 'Lodge Gift Shop',
+      description: 'Local crafts, camping supplies, WV souvenirs, and snacks',
+      fees: 'Varies',
+      seasonal: false,
+      season: 'Year-round',
+    },
+    {
+      type: 'game_room' as const,
+      name: 'Recreation Room',
+      description: 'Indoor recreation room in lodge with board games, table tennis, pool table',
+      fees: 'Free for lodge guests',
+      seasonal: false,
+    },
+  ] satisfies OtherAmenity[],
+};
